+++ types.go
@@ -4,10 +4,10 @@ import "context"
 
 type Candlestick struct {
 	Timestamp int64   `json:"time"`
-	Open      float64 `json:"open"`
-	High      float64 `json:"high"`
-	Low       float64 `json:"low"`
-	Close     float64 `json:"close"`
+	Open      float32 `json:"open"`
+	High      float32 `json:"high"`
+	Low       float32 `json:"low"`
+	Close     float32 `json:"close"`
 	Volume    float64 `json:"volume,omitempty"`
 }
